Rename: pages/StadiumPicker to pages/Register/StadiumPicker
