+++ src/Router.tsx
@@ -4,8 +4,8 @@ import Authenticated from '@pages/Authenticated';
 import Layout from '@pages/Layout';
 import Login from '@pages/Login';
 import Main from '@pages/Main';
+import StadiumPicker from '@pages/Register/StadiumPicker';
 import RequireAuth from '@pages/RequireAuth';
-import StadiumPicker from '@pages/StadiumPicker';
 import TicketRegister from '@pages/TicketRegister';
 import { useUserStore } from './store';
