+++ src/pages/Register/StadiumPicker/index.tsx
