+++ src/pages/Register/StadiumPicker/styles.ts
